Add Health and Health Bar

+++ Assets/DealDamage.cs
@@ -1,13 +1,8 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public class DealDamage : MonoBehaviour
 {
-
-    private float damage = 20;
-
-
+    private int damage = 20;
 
     private void OnTriggerEnter(Collider other)
     {
